Dir rep requires valid entries have nonzero inum
The caching bitmap allocator's rep states that the 0 entry is never in
the cache. To preserve this, the free function must not put the 0 entry
into the cache. This is enforced by the function's precondition. An
alternative plan would have been to check for 0, then drop it, but this
would have added an extra unnecessary check.

The inode allocator uses the caching allocator, so IAlloc.free requires
that the inum be nonzero. Since the inum is fetched from a directory
entry, the directory entries must not contain a 0 inum if they are
valid. Again, this could have been prevented by adding additional
checks, but the allocate functions were never allocating the 0 inode
anyway, so the new rep is already satisfied by the current code.