Actually finish lookup_ok
There's an annoying requirement that lookup requires the path to be
inside the thread's home directory. This is only because the fs_spec (in
OptimisticFSSpecs) uses an existential variable to keep track of the
result of the lookup in order to give the postcondition access;
otherwise, postconditions in fs_specs are not given access to the new
directory tree.

It should be the case that lookup is always safe. However, in any case
the result of a lookup on a non home directory file would be fairly
useless since it could change at any time. This can really only be
mitigated by layering an application protocol that is more restrictive
than directory isolation.