Define syncedmem and prove lemmas
This will be used by the log layer to keep track of blocks that are
synced in every disk in a disk sequence. The intuition is that if
`sm_ds_valid sm ds` holds, then for every address in sm that has value
`true`, that address is synced (the valuset has no tail) for evey disk
in ds.