Add syncedmem to Log.v
For each address on disk, if the synced mem has value true, then that
address is synced in all disks in the disk sequence. This is trivially
maintained by logged reads, since they don't change the disk. Logged
writes also don't change the syncedmem, since a logged write is synced
on the latest disk and doesn't affect previous ones. Only log-bypass
writes affect the syncedmem by setting the value to false.

This breaks all layers above the log, as the additional argument to
LOG.rep has not yet been propagated. This will be fixed in later
commits. The syncedmem will be exposed as a new mem for other layers to
maintain with separation logic.