Formulate BlockPtr syncedmem rep
Since the indirect block tree fans out, such that each indrep_n_tree
application causes multiple internal applications, not just one, the rep
has to be stated in such a way that it can be assembled out of the reps
for the individual pieces, hence the exists and piff.

Start pushing the change in rep through indget, indread, and inclear.