[pred_apply] should avoid instantiating evars in the mem expression
otherwise, we end up with goals like this:

  (something)%pred (list2nmem ?x)

matching some random fact in context:

  (something-else)%pred (list2nmem (some-other-list))

this fixes the proof for MemLog.dwrite_ok, and maybe others.

also add a [pred_apply_instantiate] tactic that instantiates the mem
evar during the match.  we can use this in any places where proofs
break because they used to depend on the instantiating behavior.

(for reference, the instantiating behavior came from commits 8c5504d7e,
3421b0d14, a27349b4e)